refactor: api request on utils folder
