+++ homeworks/federico-benitez/21-csr-ssr-ssg-nextjs/utils/photos.js
@@ -9,9 +9,3 @@ export const getPhotos = async (limit = 1) => {
 
   return data.map((picture) => picture.url);
 };
-
-export default async function handler(req, res) {
-  const data = getPhotos();
-
-  res.status(200).json({ data });
-}
